Split test suites
